fix: adjust the interactive modal, exports, component props, unnecessary comments
