+++ src/design/feedback/InteractiveModal/InteractiveModal.stories.tsx
@@ -5,21 +5,29 @@ import { IInteractiveModal } from "@ptypes/interactiveModal/InteractiveModalProp
 import { InteractiveModal } from ".";
 
 const story = {
-  title: "Components/Feedback/InteractiveModal",
+  title: "Components/Feedback/InteractiveModalDetailsUsers",
   component: InteractiveModal,
 };
 
 const data = {
-  missionId: "123",
+  missionId: "Pruebas",
   descriptionUse: "description",
+  descriptiosnUse: "description",
+  userID: "123",
+  username: "John Doe",
+  email: "johndoe@example.com",
+  phone: "1234567890",
+  position: "Software Engineer",
 };
 
 const dataTable = [
   {
     missionId: "123",
+    descriptionUse: "description",
   },
   {
     missionId: "456",
+    descriptionUse: "description",
   },
 ];
 
@@ -40,7 +48,8 @@ Default.args = {
   portalId: "portal",
   title: "User Information",
   infoData: data,
-  infoTitle: "Información",
+  infoTitle: "Unidades de negocio asignadas al funcionario",
+  infoText: "Cerrar",
   actionsTitle: "Acciones",
   dataTable,
 };
