+++ src/design/feedback/requestProcess/index.tsx
@@ -3,8 +3,9 @@ import { Stack, Icon, IIconAppearance, Text } from "@inubekit/inubekit";
 import { basic } from "@design/tokens";
 import { ComponentAppearance } from "@ptypes/aparences.types";
 import { IRequestSteps } from "@ptypes/feedback/requestProcess/IRequestSteps";
+import { countVerifiedRequests } from "@utils/countVerifiedRequests";
+import { verifiedErrorRequest } from "@utils/verifiedErrorRequest";
 import { StyledBar, StyledContainerBar } from "./styles";
-import { countVerifiedRequests, verifiedErrorRequest } from "./utils";
 
 interface IRequestProcess {
   appearance: IIconAppearance;
