+++ src/pages/positions/tabs/positionsTabs/forms/detailsModal/index.tsx
@@ -1,9 +1,9 @@
 import { MdOutlineRemoveRedEye } from "react-icons/md";
 import { Icon, Text } from "@inubekit/inubekit";
-import { InteractiveModal } from "@design/feedback/InteractiveModal";
 import { UseDetailsModal } from "@hooks/positions/useDetailsModal";
 import { IDetailsModal } from "@ptypes/positions/details/IDetailsModal";
 import { StyledContainerIcon } from "./styles";
+import { DetailsPositionsModal } from "./detailsPositionsModal";
 
 const DetailsModal = (props: IDetailsModal) => {
   const { data, labelsOptions } = props;
@@ -26,14 +26,12 @@ const DetailsModal = (props: IDetailsModal) => {
         )}
       </StyledContainerIcon>
       {showModal && data && (
-        <InteractiveModal
-          portalId="portal"
-          title="Detalles de cargo"
+        <DetailsPositionsModal
           infoData={data}
-          infoTitle=""
           labels={labelsOptions}
-          closeModal={handleToggleModal}
           dataTable={dataTable ?? []}
+          onClose={handleToggleModal}
+          smallScreen={screenTablet}
         />
       )}
     </>
