+++ src/utils/verifiedErrorRequest/index.ts
@@ -1,0 +1,7 @@
+import { IRequestSteps } from "@ptypes/feedback/requestProcess/IRequestSteps";
+
+const verifiedErrorRequest = (requests: IRequestSteps[]): boolean => {
+  return requests.find((request) => request.status === "error") ? true : false;
+};
+
+export { verifiedErrorRequest };
